# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PrepAI – AI Placement Mentor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List interview sessions for the current user
      responses:
        "200":
          description: List of interview sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterviewSession"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: startInterview
      tags: [interviews]
      summary: Start a new mock interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewStartInput"
      responses:
        "201":
          description: Interview session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get an interview session with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSessionWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /interviews/{id}/answer:
    post:
      operationId: submitInterviewAnswer
      tags: [interviews]
      summary: Submit an answer to the current interview question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewAnswerInput"
      responses:
        "200":
          description: AI response with follow-up or next question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewAiResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /interviews/{id}/end:
    post:
      operationId: endInterview
      tags: [interviews]
      summary: End the interview and generate a full report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /interviews/{id}/report:
    get:
      operationId: getInterviewReport
      tags: [interviews]
      summary: Get the report for a completed interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Companies
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Dashboard
...
    InterviewSession:
      type: object
      required: [id, userId, role, difficulty, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        role:
          type: string
        difficulty:
          type: string
        status:
          type: string
        overallScore:
          type: ["number", "null"]
        questionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InterviewSessionWithMessages:
      type: object
      required: [id, userId, role, difficulty, status, messages, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        role:
          type: string
        difficulty:
          type: string
        status:
          type: string
        overallScore:
          type: ["number", "null"]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/InterviewMessage"
        createdAt:
          type: string
          format: date-time
    InterviewMessage:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
        content:
          type: string
        score:
          type: ["number", "null"]
        feedback:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    InterviewStartInput:
      type: object
      required: [role, difficulty]
      properties:
        role:
          type: string
          enum: [Frontend, Backend, Java, MERN, SDE, Full Stack]
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
    InterviewAnswerInput:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
          minLength: 1
    InterviewAiResponse:
      type: object
      required: [message, type]
      properties:
        message:
          type: string
        type:
          type: string
          enum: [cross_question, next_question, end]
        score:
          type: ["number", "null"]
        feedback:
          type: ["string", "null"]
        questionNumber:
          type: integer
    InterviewReport:
      type: object
      required: [sessionId, overallScore, confidence, communication, technicalSkills, strongAreas, weakAreas, recommendedTopics]
      properties:
        sessionId:
          type: integer
        overallScore:
          type: number
        confidence:
          type: number
        communication:
          type: number
        technicalSkills:
          type: number
        strongAreas:
          type: array
          items:
            type: string
        weakAreas:
          type: array
          items:
            type: string
        recommendedTopics:
          type: array
          items:
            type: string

    # Companies
...
    UserProfile:
      type: object
      required: [userId, skills, targetCompany, targetRole, preparationProgress]
      properties:
        userId:
          type: string
        skills:
          type: array
          items:
            type: string
        projects:
          type: array
          items:
            type: string
        targetCompany:
          type: ["string", "null"]
        targetRole:
          type: ["string", "null"]
        preparationProgress:
          type: number
        bio:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        graduationYear:
          type: ["integer", "null"]
    UserProfileUpdate:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
        projects:
          type: array
          items:
            type: string
        targetCompany:
          type: string
        targetRole:
          type: string
        bio:
          type: string
        college:
          type: string
        graduationYear:
          type: integer