# lib/api-zod/src/generated/types/interviewSessionWithMessages.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PrepAI – AI Placement Mentor API
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewMessage } from './interviewMessage';

export interface InterviewSessionWithMessages {
  id: number;
  userId: string;
  role: string;
  difficulty: string;
  status: string;
  /** @nullable */
  overallScore?: number | null;
  messages: InterviewMessage[];
  createdAt: Date;
}